% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRM.R
\name{CCRM}
\alias{CCRM}
\title{Constrained Center and Range Method for Interval-Valued Data Regression}
\usage{
CCRM(formula = formula, data = data)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}, a symbolic description of the model to be fitted.}

\item{data}{an data frame containing the variables in the model.}
}
\value{
\item{coefficients.Center}{Coefficients for the Center variable.}

\item{coefficeints.Range}{Coefficients for the Range variable.}

\item{fitted.values}{The fitted values for the lower and upper interval bound.}

\item{residuals}{The residuals for the lower and upper interval bound.}
}
\description{
\code{CCRM()} is used to fit a linear regression model based on the inequality constraints over the range variables.(Lima Neto and De Carvalho, 2010)
}
\details{
The disadvantages of the CRM model is that the predicted valu of the range model may be negative. Similar to Center and Range Method(CRM), but adds constraint condition that all estimative of the parameters of the range's model are positive.(based on inequality constraints)
There is no constraints over the parameters estimates for the center point regression equation.
}
\note{
In dataset, a pair of the interval variables should always be composed in order from lower to upper bound. In order to apply this function, the data should be composed as follows:
\tabular{cccccc}{
\eqn{y_L}     \tab  \eqn{y_U}     \tab  \eqn{x1_L}    \tab  \eqn{x1_U}    \tab  \eqn{x2_L}    \tab  \eqn{x2_U}\cr
\eqn{y_L1}  \tab  \eqn{y_U1}  \tab  \eqn{x_L11} \tab  \eqn{x_U11} \tab  \eqn{x_L12} \tab  \eqn{x_U12}\cr
\eqn{y_L2}  \tab  \eqn{y_U2}  \tab  \eqn{x_L21} \tab  \eqn{x_U21} \tab  \eqn{x_L22} \tab  \eqn{x_U22}\cr
\eqn{y_L3}  \tab  \eqn{y_U3}  \tab  \eqn{x_L31} \tab  \eqn{x_U31} \tab  \eqn{x_L32} \tab  \eqn{x_U32}\cr
\eqn{y_L4}  \tab  \eqn{y_U4}  \tab  \eqn{x_L41} \tab  \eqn{x_U41} \tab  \eqn{x_L42} \tab  \eqn{x_U42}\cr
\eqn{y_L5}  \tab  \eqn{y_U5}  \tab  \eqn{x_L51} \tab  \eqn{x_U51} \tab  \eqn{x_L52} \tab  \eqn{x_U52}\cr
}

The upper limit value of the variable should be unconditionally greater than the lower limit value. Otherwise, it will be output as \code{NA} or \code{NAN}, and the value can not be generated.
}
\examples{
data(example3)
m1 <- CCRM(cbind(Sepal.Length_L, Sepal.Length_U) ~ Sepal.Width_L + Sepal.Width_U + Petal.Length_L + Petal.Length_U + Petal.Length_L + Petal.Length_U, data = example3)
m1
m1$coefficients
}
\references{
Lima Neto, E.A. and De Carvalho, F.A.T(2010), Constrained linear regression models for symbolic interval-valued variables \emph{Computational Statistics and Data Analysis, 54}, 333-347
}
\seealso{
\code{\link{RMSE}} \code{\link{symbolic.r}}
}

