% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRM.R
\name{CRM}
\alias{CRM}
\title{Center and Range Method for Interval-Valued Data Regression}
\usage{
CRM(formula = formula, data = data)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}, a symbolic description of the model to be fitted.}

\item{data}{an data frame containing the variables in the model.}
}
\value{
\item{coefficients.Center}{Coefficients for the Center variable.}

\item{coefficeints.Range}{Coefficients for the Range variable.}

\item{fitted.values}{The fitted values for the lower and upper interval bound.}

\item{residuals}{The residuals for the lower and upper interval bound.}
}
\description{
\code{CRM()} is used to fit a linear regression model based on the Center and Range Method(Lima Neto and De Carvalho, 2008)
}
\details{
The basic idea is to estimate models independently for each of center points and range points of the interval variables. The coefficients of the model are estimated by least square method.
}
\note{
In dataset, a pair of the interval variables should always be composed in order from lower to upper bound. In order to apply this function, the data should be composed as follows:
\tabular{cccccc}{
\eqn{y_L}     \tab  \eqn{y_U}     \tab  \eqn{x1_L}    \tab  \eqn{x1_U}    \tab  \eqn{x2_L}    \tab  \eqn{x2_U}\cr
\eqn{y_L1}  \tab  \eqn{y_U1}  \tab  \eqn{x_L11} \tab  \eqn{x_U11} \tab  \eqn{x_L12} \tab  \eqn{x_U12}\cr
\eqn{y_L2}  \tab  \eqn{y_U2}  \tab  \eqn{x_L21} \tab  \eqn{x_U21} \tab  \eqn{x_L22} \tab  \eqn{x_U22}\cr
\eqn{y_L3}  \tab  \eqn{y_U3}  \tab  \eqn{x_L31} \tab  \eqn{x_U31} \tab  \eqn{x_L32} \tab  \eqn{x_U32}\cr
\eqn{y_L4}  \tab  \eqn{y_U4}  \tab  \eqn{x_L41} \tab  \eqn{x_U41} \tab  \eqn{x_L42} \tab  \eqn{x_U42}\cr
\eqn{y_L5}  \tab  \eqn{y_U5}  \tab  \eqn{x_L51} \tab  \eqn{x_U51} \tab  \eqn{x_L52} \tab  \eqn{x_U52}\cr
}

The upper limit value of the variable should be unconditionally greater than the lower limit value. Otherwise, it will be output as \code{NA} or \code{NAN}, and the value can not be generated.
}
\examples{
set.seed(2017)
x1_L = rnorm(30, 3, 0.01) - rnorm(30, 0, 0.01)
x1_U = rnorm(30, 3, 0.01) + rnorm(30, 3, 0.01)
x2_L = runif(30, 1.5, 3) - runif(30, 0, 1)
x2_U = runif(30, 1.5, 3) + runif(30, 1, 2)
y_L = x1_L + x2_L
y_U = x1_U + x2_U
temp <- as.data.frame(cbind(y_L, y_U, x1_L, x1_U, x2_L, x2_U))
m1 <- CRM(cbind(y_L, y_U) ~ x1_L + x1_U + x2_L + x2_U, data = temp)
m1
}
\references{
Billard, L. and Diday, E.(2000), Regression analysis for interval-valued data. \emph{Data Analysis, Classification, and Related Methods.} Springer-Verlag, Berlin, 369-374

Lima Neto, E.A. and De Carvalho, F.A.T(2008), Centre and range method to fitting a linear regression model on symbolic interval data. \emph{Computational Statistics and Data Analysis, 52}, 1500-1515
}
\seealso{
\code{\link{RMSE}} \code{\link{symbolic.r}}
}

