% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example1.R
\docType{data}
\name{example1}
\alias{example1}
\title{Interval-Valued data example}
\format{An object of class \code{data.frame} with 15 rows and 6 columns.}
\source{
\url{http://www.stat.uga.edu/data-tables-billard-l-and-diday-e-2007}
}
\usage{
example1
}
\description{
Blood pressure data
}
\examples{
data(example1)
}
\references{
Billard L. and Diday E.(2006), Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}

