% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example2.R
\docType{data}
\name{example2}
\alias{example2}
\title{Interval-Valued data example}
\format{An object of class \code{data.frame} with 23 rows and 7 columns.}
\source{
\url{http://www.stat.uga.edu/data-tables-billard-l-and-diday-e-2007}
}
\usage{
example2
}
\description{
Mushroom data
}
\examples{
data(example2)
}
\references{
Billard L. and Diday E.(2006), Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}

