% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example3.R
\docType{data}
\name{example3}
\alias{example3}
\title{Interval-Valued data example}
\format{An object of class \code{data.frame} with 30 rows and 9 columns.}
\source{
\url{http://www.stat.uga.edu/data-tables-billard-l-and-diday-e-2007}
}
\usage{
example3
}
\description{
Iris data
}
\examples{
data(example3)
}
\references{
Billard L. and Diday E.(2006), Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}

