% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example4.R
\docType{data}
\name{example4}
\alias{example4}
\title{Interval-Valued data example}
\format{An object of class \code{data.frame} with 8 rows and 9 columns.}
\source{
\url{http://www.stat.uga.edu/data-tables-billard-l-and-diday-e-2007}
}
\usage{
example4
}
\description{
Cars data
}
\examples{
data(example4)
}
\references{
Billard L. and Diday E.(2006), Symbolic data analysis: Conceptual statistics and data mining. Wiley, Chichester.
}
\keyword{datasets}

