% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcmtn.R
\name{imcmtn}
\alias{imcmtn}
\title{MCM for Interval-Valued Data Regression Using Truncated normal distribution}
\usage{
imcmtn(formula = formula, data = data, b = 100)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}, a symbolic description of the model to be fitted.}

\item{data}{an data frame containing the variables in the model.}

\item{b}{number of resampling (default : 100)}
}
\value{
\item{resampling.coefficients}{B coefficient vectors obtained by resampling.}

\item{coefficients}{Average of B coefficients obtained by resampling, standard error and p-value.}

\item{fitted.values}{The fitted values for the lower and upper interval bound.}

\item{residuals}{The residuals for the lower and upper interval bound.}
}
\description{
\code{imcmtn()} is used to fit a linear regression model based on the Monte Carlo Method using truncated normal distribution.
}
\details{
Similar to \code{imcmuni()}, but uses a truncated normal distribution rather than a uniform distribution.
}
\note{
In dataset, a pair of the interval variables should always be composed in order from lower to upper bound. In order to apply this function, the data should be composed as follows:
\tabular{cccccc}{
\eqn{y_L}     \tab  \eqn{y_U}     \tab  \eqn{x1_L}    \tab  \eqn{x1_U}    \tab  \eqn{x2_L}    \tab  \eqn{x2_U}\cr
\eqn{y_L1}  \tab  \eqn{y_U1}  \tab  \eqn{x_L11} \tab  \eqn{x_U11} \tab  \eqn{x_L12} \tab  \eqn{x_U12}\cr
\eqn{y_L2}  \tab  \eqn{y_U2}  \tab  \eqn{x_L21} \tab  \eqn{x_U21} \tab  \eqn{x_L22} \tab  \eqn{x_U22}\cr
\eqn{y_L3}  \tab  \eqn{y_U3}  \tab  \eqn{x_L31} \tab  \eqn{x_U31} \tab  \eqn{x_L32} \tab  \eqn{x_U32}\cr
\eqn{y_L4}  \tab  \eqn{y_U4}  \tab  \eqn{x_L41} \tab  \eqn{x_U41} \tab  \eqn{x_L42} \tab  \eqn{x_U42}\cr
\eqn{y_L5}  \tab  \eqn{y_U5}  \tab  \eqn{x_L51} \tab  \eqn{x_U51} \tab  \eqn{x_L52} \tab  \eqn{x_U52}\cr
}

The upper limit value of the variable should be unconditionally greater than the lower limit value. Otherwise, it will be output as \code{NA} or \code{NAN}, and the value can not be generated.
}
\examples{
set.seed(2017)
x1_L = rnorm(30, 3, 0.01) - rnorm(30, 0, 0.01)
x1_U = rnorm(30, 3, 0.01) + rnorm(30, 3, 0.01)
x2_L = runif(30, 1.5, 3) - runif(30, 0, 1)
x2_U = runif(30, 1.5, 3) + runif(30, 1, 2)
y_L = x1_L + x2_L
y_U = x1_U + x2_U
temp <- as.data.frame(cbind(y_L, y_U, x1_L, x1_U, x2_L, x2_U))
m1 <- imcmtn(cbind(y_L, y_U) ~ x1_L + x1_U + x2_L + x2_U, data = temp, b = 100)
m1

}
\references{
Ahn, J., Peng, M., Park, C., Jeon, Y.(2012), A Resampling Approach for Interval-Valued Data Regression. \emph{Statistical Analysis and Data Mining, 5, 336-348}

Lim, S., Kang, K.(2016), On Statistical Analysis of Interval-Valued Data
}
\seealso{
\code{\link[truncnorm]{rtruncnorm}} \code{\link{RMSE}} \code{\link{symbolic.r}}
}

