% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcmuni.R
\name{imcmuni}
\alias{imcmuni}
\title{MCM for Interval-Valued Data Regression Using Uniform distribution}
\usage{
imcmuni(formula = formula, data = data, b = 100)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}, a symbolic description of the model to be fitted.}

\item{data}{an data frame containing the variables in the model.}

\item{b}{number of resampling (default : 100)}
}
\value{
\item{resampling.coefficients}{B coefficient vectors obtained by resampling.}

\item{coefficients}{Average of B coefficients obtained by resampling, standard error and p-value.}

\item{fitted.values}{The fitted values for the lower and upper interval bound.}

\item{residuals}{The residuals for the lower and upper interval bound.}
}
\description{
\code{imcmuni()} is used to fit a linear regression model based on the Monte Carlo Method using uniform distribution.
}
\details{
Ahn et al.(2012) propose a regression approach for interval-valued data based on resampling. So this method is to resample by randomly selecting a single-valued point within each observed intervals.
Then, fit a classical linear regression model on each single-valued points, and calculate the average of regression coefficients over the models. The use of the resampling approach method, called Monte Carlo method (MCM), has the advantage of estimating on sample distribution approximately, and statistical inference is possible using this.
}
\note{
In dataset, a pair of the interval variables should always be composed in order from lower to upper bound. In order to apply this function, the data should be composed as follows:
\tabular{cccccc}{
\eqn{y_L}     \tab  \eqn{y_U}     \tab  \eqn{x1_L}    \tab  \eqn{x1_U}    \tab  \eqn{x2_L}    \tab  \eqn{x2_U}\cr
\eqn{y_L1}  \tab  \eqn{y_U1}  \tab  \eqn{x_L11} \tab  \eqn{x_U11} \tab  \eqn{x_L12} \tab  \eqn{x_U12}\cr
\eqn{y_L2}  \tab  \eqn{y_U2}  \tab  \eqn{x_L21} \tab  \eqn{x_U21} \tab  \eqn{x_L22} \tab  \eqn{x_U22}\cr
\eqn{y_L3}  \tab  \eqn{y_U3}  \tab  \eqn{x_L31} \tab  \eqn{x_U31} \tab  \eqn{x_L32} \tab  \eqn{x_U32}\cr
\eqn{y_L4}  \tab  \eqn{y_U4}  \tab  \eqn{x_L41} \tab  \eqn{x_U41} \tab  \eqn{x_L42} \tab  \eqn{x_U42}\cr
\eqn{y_L5}  \tab  \eqn{y_U5}  \tab  \eqn{x_L51} \tab  \eqn{x_U51} \tab  \eqn{x_L52} \tab  \eqn{x_U52}\cr
}

The upper limit value of the variable should be unconditionally greater than the lower limit value. Otherwise, it will be output as \code{NA} or \code{NAN}, and the value can not be generated.
}
\examples{
set.seed(2017)
x1_L = rnorm(30, 3, 0.01) - rnorm(30, 0, 0.01)
x1_U = rnorm(30, 3, 0.01) + rnorm(30, 3, 0.01)
x2_L = runif(30, 1.5, 3) - runif(30, 0, 1)
x2_U = runif(30, 1.5, 3) + runif(30, 1, 2)
y_L = x1_L + x2_L
y_U = x1_U + x2_U
temp <- as.data.frame(cbind(y_L, y_U, x1_L, x1_U, x2_L, x2_U))
m1 <- imcmuni(cbind(y_L, y_U) ~ x1_L + x1_U + x2_L + x2_U, data = temp, b = 100)
m1

}
\references{
Ahn, J., Peng, M., Park, C., Jeon, Y.(2012), A Resampling Approach for Interval-Valued Data Regression. \emph{Statistical Analysis and Data Mining, 5}, 336-348
}
\seealso{
\code{\link{RMSE}} \code{\link{symbolic.r}}
}

