% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.beta.R
\name{int.beta}
\alias{int.beta}
\title{Perform interval censored-regression assuming data are bounded continuous (assuming beta distribution)}
\usage{
int.beta(formula, data, min = 0, max = 1, link = "logit", ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class).
Must be of form: cbind(lower, upper) ~ ... where lower is the lower bound and upper
is the upper bound.}

\item{data}{a data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model}

\item{min}{lower bound for data, 0 default}

\item{max}{upper bound for data, 1 default}

\item{link}{link function, either logit (default) or probit}
}
\value{
A fitted model object of class mle2, Call model.summary() function for inference.
}
\description{
Perform interval censored-regression assuming data are bounded continuous (assuming beta distribution)
}
