% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.binomial.R
\name{int.binomial}
\alias{int.binomial}
\title{Perform interval censored-regression assuming data are binomial}
\usage{
int.binomial(formula, data, max, link = "logit", ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class).
Must be of form: cbind(lower, upper) ~ ... where lower is the lower bound and upper
is the upper bound.}

\item{data}{a data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model}

\item{max}{number of items summed to create the original scores}

\item{link}{link function, either logit (default) or probit}
}
\value{
A fitted model object of class mle2, Call model.summary() function for inference.
}
\description{
Perform interval censored-regression assuming data are binomial
}
