% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.helpers.R
\name{model.summary}
\alias{model.summary}
\title{Model summary}
\usage{
model.summary(object)
}
\arguments{
\item{object}{fitted object}
}
\value{
Print summary of fitted model.
}
\description{
Model summary
}
