% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCM.R
\name{SCM}
\alias{SCM}
\title{Symbolic Covariance Method}
\usage{
SCM(formula = formula, data = data)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}, a symbolic description of the model to be fitted.}

\item{data}{an data frame containing the variables in the model.}
}
\value{
\item{symbolic.covariance.Sxx}{Symbolic sample variance-covariance matrix between response variable Y and predictor variables X.}

\item{symbolic.covariance.Sxy}{Symbolic sample covariance vector between response variable Y and predictor variables X}

\item{coefficients}{regression coefficients}

\item{fitted.values}{The fitted values for the lower and upper interval bound.}

\item{residuals}{The residuals for the lower and upper interval bound.}
}
\description{
\code{SCM()} is used to fit a linear regression model based on symbolic covariance matrix(Xu, 2010).
}
\details{
The SCM proposed by Xu(2010) is a method of estimating a regression coefficient using a symbolic covariance matrix. In SCM, the centralized linear regression model is used(model with centered variables).
The regression coefficient is estimated by least squares method, but it is used by symbolic covariance matrix. Because the process of calculating the symbolic sample covariance uses the lower and upper limits of each variable, the SCM reflects the variability of the interval.
}
\note{
In dataset, a pair of the interval variables should always be composed in order from lower to upper bound. In order to apply this function, the data should be composed as follows:
\tabular{cccccc}{
\eqn{y_L}     \tab  \eqn{y_U}     \tab  \eqn{x1_L}    \tab  \eqn{x1_U}    \tab  \eqn{x2_L}    \tab  \eqn{x2_U}\cr
\eqn{y_L1}  \tab  \eqn{y_U1}  \tab  \eqn{x_L11} \tab  \eqn{x_U11} \tab  \eqn{x_L12} \tab  \eqn{x_U12}\cr
\eqn{y_L2}  \tab  \eqn{y_U2}  \tab  \eqn{x_L21} \tab  \eqn{x_U21} \tab  \eqn{x_L22} \tab  \eqn{x_U22}\cr
\eqn{y_L3}  \tab  \eqn{y_U3}  \tab  \eqn{x_L31} \tab  \eqn{x_U31} \tab  \eqn{x_L32} \tab  \eqn{x_U32}\cr
\eqn{y_L4}  \tab  \eqn{y_U4}  \tab  \eqn{x_L41} \tab  \eqn{x_U41} \tab  \eqn{x_L42} \tab  \eqn{x_U42}\cr
\eqn{y_L5}  \tab  \eqn{y_U5}  \tab  \eqn{x_L51} \tab  \eqn{x_U51} \tab  \eqn{x_L52} \tab  \eqn{x_U52}\cr
}

The upper limit value of the variable should be unconditionally greater than the lower limit value. Otherwise, it will be output as \code{NA} or \code{NAN}, and the value can not be generated.
}
\examples{
data(example3)
m1 <- SCM(cbind(Sepal.Length_L, Sepal.Length_U) ~ Sepal.Width_L + Sepal.Width_U + Petal.Length_L + Petal.Length_U + Petal.Length_L + Petal.Length_U, data = example3)
m1
m1$coefficients
}
\references{
Xu, W.(2010), Symbolic Data Analysis: Interval-Valued Data Regression
}
\seealso{
\code{\link{RMSE}} \code{\link{symbolic.r}}
}

