% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic.r
\name{symbolic.r}
\alias{symbolic.r}
\title{The Symbolic Correlation}
\usage{
symbolic.r(model)
}
\arguments{
\item{model}{\code{\link[imcm]{imcmuni}} or \code{\link[imcm]{imcmtn}} object, etc..}
}
\description{
The symbolic correlation coefficient, \eqn{r}, proposed by Billard(2007, 2008) and applied to the regression problem by Xu(2010), measures the correlation between the predicted values and the observed values.
}
\examples{
set.seed(2017)
x1_L = rnorm(30, 3, 0.01) - rnorm(30, 0, 0.01)
x1_U = rnorm(30, 3, 0.01) + rnorm(30, 3, 0.01)
x2_L = runif(30, 1.5, 3) - runif(30, 0, 1)
x2_U = runif(30, 1.5, 3) + runif(30, 1, 2)
y_L = x1_L + x2_L
y_U = x1_U + x2_U
temp <- as.data.frame(cbind(y_L, y_U, x1_L, x1_U, x2_L, x2_U))
m1 <- imcmtn(cbind(y_L, y_U) ~ x1_L + x1_U + x2_L + x2_U, data = temp, b = 100)
symbolic.r(m1)
}
\references{
Billard(2007, 2008). Sample covariance functions for complex quantitative data

Xu(2010), Symbolic Data Analysis: Interval-Valued Data Regression
}

